+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+- Invert verbosity control by removing `-q/--quiet` and adding `-v/--verbose`.
+  Verbose output is now opt-in, not opt-out.
+
 # 0.0.2
 
 - Add pre-commit-hooks config, allowing use with pre-commit